feat: Populate 'firm' field in getVendorById endpoint
This commit modifies the `getVendorById` endpoint in the `vendorController.js` file. It adds the `.populate('firm')` method to the `Vendor.findById` query, allowing the 'firm' field to be populated with the associated firm data. This change ensures that when retrieving a vendor by ID, the response includes the firm information as well.