feat: Add deleteFirmById endpoint
This commit adds the `deleteFirmById` endpoint to the `firmController.js` file. It allows the deletion of a firm by its ID. If the firm is not found, a 404 status with a corresponding error message is returned. If there is a server error, a 500 status with a server error message is returned.